pyproject.toml
[tool.poetry]
name = "eq-runner-mock-sds"
version = "0.1.0"
description = "ONS Digital eQ Runner Mock SDS"
authors = ["http://onsdigital.github.io"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
sdc-cryptography = "^1.1.2"
uvicorn = "^0.23.0"
gunicorn = "^20.1.0"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
